M4: Riddle: Fix interface eventually showing during credits (#16070), small cleanup

+++ engines/m4/riddle/rooms/section4/room493.h
@@ -34,7 +34,7 @@ private:
 	int _panning = 0;
 
 	static void escape_key_pressed(void *, void *);
-	void setTrigger(int trigger);
+	void setTrigger(int16 trigger);
 	void crunch();
 
 public:
